fastapi/requirements.txt
fastapi==0.68.0
pillow==9.0.1
python-multipart==0.0.5 
aiofiles==0.4.0
streamlit==0.86.0 
requests==2.24.0
requests-toolbelt==0.9.1
yacs
opencv-python
PyYAML
cython-bbox
scipy
progress
motmetrics
matplotlib
lap
openpyxl
Pillow
tensorboardX
fvcore
protobuf==3.19.0
mmcv
timm
